+++ example/lib/pages/cache_page.dart
@@ -9,6 +9,7 @@ class CachePage extends StatefulWidget {
 
 class _CachePageState extends State<CachePage> {
   late BetterPlayerController _betterPlayerController;
+  late BetterPlayerDataSource _betterPlayerDataSource;
 
   @override
   void initState() {
@@ -17,13 +18,16 @@ class _CachePageState extends State<CachePage> {
       aspectRatio: 16 / 9,
       fit: BoxFit.contain,
     );
-    BetterPlayerDataSource dataSource = BetterPlayerDataSource(
+    _betterPlayerDataSource = BetterPlayerDataSource(
       BetterPlayerDataSourceType.network,
-      Constants.forBiggerBlazesUrl,
-      cacheConfiguration: BetterPlayerCacheConfiguration(useCache: true),
+      Constants.elephantDreamVideoUrl,
+      cacheConfiguration: BetterPlayerCacheConfiguration(
+          useCache: true,
+          preCacheSize: 10 * 1024 * 1024,
+          maxCacheSize: 10 * 1024 * 1024,
+          maxCacheFileSize: 10 * 1024 * 1024),
     );
     _betterPlayerController = BetterPlayerController(betterPlayerConfiguration);
-    _betterPlayerController.setupDataSource(dataSource);
     super.initState();
   }
 
@@ -51,11 +55,29 @@ class _CachePageState extends State<CachePage> {
             child: BetterPlayer(controller: _betterPlayerController),
           ),
           TextButton(
+            child: Text("Start pre cache"),
+            onPressed: () {
+              _betterPlayerController.preCache(_betterPlayerDataSource);
+            },
+          ),
+          TextButton(
+            child: Text("Stop pre cache"),
+            onPressed: () {
+              _betterPlayerController.stopPreCache(_betterPlayerDataSource);
+            },
+          ),
+          TextButton(
+            child: Text("Play video"),
+            onPressed: () {
+              _betterPlayerController.setupDataSource(_betterPlayerDataSource);
+            },
+          ),
+          TextButton(
             child: Text("Clear cache"),
             onPressed: () {
               _betterPlayerController.clearCache();
             },
-          )
+          ),
         ],
       ),
     );