+++ lib/src/video_player/video_player_platform_interface.dart
@@ -74,6 +74,16 @@ abstract class VideoPlayerPlatform {
     throw UnimplementedError('create() has not been implemented.');
   }
 
+  /// Pre-caches a video.
+  Future<void> preCache(DataSource dataSource, int preCacheSize) {
+    throw UnimplementedError('preCache() has not been implemented.');
+  }
+
+  /// Pre-caches a video.
+  Future<void> stopPreCache(String url) {
+    throw UnimplementedError('stopPreCache() has not been implemented.');
+  }
+
   /// Set data source of video.
   Future<void> setDataSource(int? textureId, DataSource dataSource) {
     throw UnimplementedError('setDataSource() has not been implemented.');
@@ -156,7 +166,7 @@ abstract class VideoPlayerPlatform {
     throw UnimplementedError('setMixWithOthers() has not been implemented.');
   }
 
-  Future<void> clearCache(int? textureId) {
+  Future<void> clearCache() {
     throw UnimplementedError('clearCache() has not been implemented.');
   }